feat: Create the SelectSubWarehouse component
