+++ src/layouts/Header.vue
@@ -1,24 +1,17 @@
 <script setup lang="ts">
 import { ref, computed } from "vue";
 import { useAuthStore } from "@/modules/Auth/stores"
-import { useSubStore } from "@/modules/Store/stores/"
 import { useSidebar } from "@/composables/useSidebar"
 import Logout from "@/modules/Auth/components/Logout.vue"
 //import LoginIcon from "@/icons/LoginIcon.vue"
 import HomeIcon from "@/icons/HomeIcon.vue"
 import SelectTheme from "./SelectTheme.vue"
-import AppSelect from "@/components/AppSelect.vue";
+import SelectSubWarehouse from "@/modules/Store/components/SubWarehouse/SelectSubWarehouse.vue"
 //import IconLogo from "@/icons/IconLogo.vue"
 
 const { isOpen, isClose } = useSidebar()
 const dropdownOpen = ref(false)
 const store = computed(() => useAuthStore())
-//const subStore = computed(() => useSubStore())
-const subStore = useSubStore()
-
-
-subStore.getSubStoreIni()
-
 </script>
 
 <template>
@@ -100,20 +93,7 @@ subStore.getSubStoreIni()
     </div>
 
     <div class="flex items-center">
-      <div class="relative">
-        <select v-model="subStore.uuid" @change="subStore.handleChange">
-          <option key="" value="" disabled selected>Seleccione...</option>
-          <option
-            v-for="option in [{id: 'sub-store-one', name: 'Sub Almacén Uno'}, {id: 'sub-store-two', name: 'Sub Almacén Dos'}, {id: 'sub-store-three', name: 'Sub Almacén Tres'}]"
-            :value="option.id"
-            :selected="option.id === subStore.uuid"    
-          >{{ option.name }}</option>
-        </select>
-      </div>
-
-
-
-
+      <SelectSubWarehouse />
 
       <button class="flex focus:outline-none">
         <svg
